+++ source/Examples/Resources/DnsClientNrptRule/3-DnsClientNrptRule_Punycode_Config.ps1
@@ -19,15 +19,7 @@
 
 <#
     .DESCRIPTION
-        Sets a Dns Client NRPT rule to send Punycode DNS queries using a conditional dns forwarder for a specific namespace.
-    .PARAMETER Name
-        Specifies the name which uniquely identifies a rule.
-    .PARAMETER NameEncoding
-        Specifies the encoding format for host names in the DNS query.
-    .PARAMETER NameServers
-        Specifies the DNS servers to which the DNS query is sent when DirectAccess is disabled.
-    .PARAMETER Namespace
-        Specifies the DNS namespace.
+        Sets a DNS Client NRPT rule named 'Punycode' to send Punycode DNS queries using a conditional DNS forwarder for a specific namespace(contoso.com).
 #>
 Configuration DnsClientNrptRule_Punycode_Config
 {